+++ test/expected/transform/js/bar/c/c2.js
@@ -1,8 +1,8 @@
-var a = require('./../../foo/a/index');
+var a = require('../../foo/a/index');
 
-var b1 = require('./../../foo/b/b1');
+var b1 = require('../../foo/b/b1');
 
-var b2 = require('./../../foo/b/b2');
+var b2 = require('../../foo/b/b2');
 
 var a1 = new a.A1();
 var a2 = new a.A2();