+++ test/expected/transform/ts/foo/a/A1.ts
@@ -1,4 +1,6 @@
-
+///ts:ref=A2
+/// <reference path="./A2.ts"/> ///ts:ref:generated
+  
 class A1 {
     a1() { return "a1" }
 }