+++ test/transform/ts/bar/c/c2.ts
@@ -1,9 +1,9 @@
 ///ts:import=a
-import a = require('./../../foo/a/index'); ///ts:import:generated
+import a = require('../../foo/a/index'); ///ts:import:generated
 ///ts:import=b1
-import b1 = require('./../../foo/b/b1'); ///ts:import:generated
+import b1 = require('../../foo/b/b1'); ///ts:import:generated
 ///ts:import=b2
-import b2 = require('./../../foo/b/b2'); ///ts:import:generated
+import b2 = require('../../foo/b/b2'); ///ts:import:generated
 
 // USE CLASSES FROM A/B: